Switch to mu4e from notmuch
Though notmuch seems to be the superior choice for me because of its
ideology, mu4e is just more convenient and has way better support for
switching or intelligently choosing identities.

Since I'm working on 2-3 machines simultaneously it's necessary for me
to have a synchronized "state" of my emails at any point in time. This
is way harder to achieve with notmuch, even using muchsync.