Create custom (non-Doom) emacs module
This marks the first step in trying to build my "own" emacs config
from scratch, not using Doom, Spacemacs, or anything like that.

So far I'm not even sure what I'd need and may well abort this
experiment after some amount of tedium, but I feel like this is going
to be a great experience.