Disable evil-mode and friends for now
This is just to see whether I can learn to live with the default Emacs
keybindings. It's actually quite exciting to try this as something
new, and not use vim emulation everywhere. Who knows, maybe it'll end
up being efficient enough while not masking a lot of the underlying
and intended user interfaces.