Zap rofi, use dmenu instead
I like Rofi, but a lot of programs have issues when started this way
as opposed to via dmenu.  For instance, Thunderbird, Firefox, and
Chrome all crash when trying to download files.  The errors have to do
with the environment and probably GTK, but it's very hard to
investigate and not worth it for me now.