Stop muchsync from doing `notmuch new` locally
I'm using the Fcc header to move sent mail to the respective folders
for my accounts, so that I can see it on the server as well as in
other mail clients.  But there were 2 problems:

1) After sending a mail, notmuch.el's default is to use `notmuch
insert`.  But this way AFAIU I'd have to specify the tags myself, and
redundantly at that, as it circumvents the post-new machinery on the
server.

2) The real problem: muchsync needs to be run with the --nonew option
to disable local "notmuch new".  That in itself wouldn't hurt, but
it's only the default notmuch configuration, as muchsync sets it up
itself on first configuration.  Due to the setup there is actually no
need to have a notmuch configuration with initial tagging on the
client machines, as the server does it all.  Now it hopefully should
work for sent mail as well, as they're moved via Fcc, then handled by
the server.  It should only cost a slight delay until they appear
locally, as they now need to be processed by the muchsync service
once.