Re-enable automatic insert mode on 'magit-commit'
Fixed the problem of the cursor color permanently changing to the
insert state color by checking for the name of the (current-buffer).
When the buffer for `magit-commit' opens, the current buffer in
`server-after-make-frame-hook' is named " *server*".  I now check for
that and in this case do not change my cursor colors.