Give up on gnus (for now)
I evaluated three different programs/frontends for email (in case of
gnus it's of course not only about email):

- notmuch + mbsync (on a server) + muchsync
- mu4e (a couple of years ago in Doom Emacs)
- gnus

Since I'm working on multiple machines, I want my setup to be as
portable as possible, meaning that I should just be able to start
Emacs on any one of these machines and see the correct state of my
emails, be it work or private.  Things I learned:

Notmuch with muchsync to synchronize the tag database and sent mail
works, but seems to be brittle.  On the server, I need to have both
preNew and postNew hooks (for mbsync and initial tagging,
respectively).  On the clients, muchsync --nonew is usually what I
want (and the only variant that doesn't throw any errors, though it
always seems to work still).  Sometimes I need to run it without the
--nonew option, namely when getting "returned" sent mail.  Plus
sometimes (after not using one machine for a while usually), I'd see
up to 20 old emails that I had already read marked as unread in my
inbox, which I don't understand and is thus quite annoying.

Gnus is very mighty and customizable, and I think I could "get there",
but then what annoys me is that I still have to use some external
synchronization mechanism for the .newsrc.eld file (and others).

So that leaves mu4e, which I will probably give another try soon.