Try out a new look
The backgrounds of the Emacs modeline, rofi, and the other i3
components should be in sync.  I'll try a new "look" where the
backgrounds of "active" things are green instead of just black.  In
particular the modeline of the active buffer is now more easily
distinguishable from that of other buffers.