emacs: use org-mode for literate configuration
I've done it in the past and gone back to a pure ELisp configuration.
Now that I'm thinking of writing some small blog entries and using
org-mode for this, it only seems natural to go back to a literate
config.  I could then export my configuration to a nice HTML web page
via 'ox-publish.