emacs: tame some magit buffers
COMMIT_EDITMSG does not use #'display-buffer, ergo is not tameable by
shackle.  The trick to getting a usable configuration is not matching
on "^magit.*" buffers, specifically leaving the magit-diff buffer
alone.  Otherwise the focus handling seems to break entirely.

Now it seems like since the magit-diff window is (in my case) always
created after having done magit-status, it reuses the same window,
implicitly following the shackle rules for magit-status-mode.  Then
it's split vertically to display the COMMIT_EDITMSG buffer in the
left-hand side.