emacs: zap evil-mode specific cursor colors for now
This should probably become a separate package, if I ever get to do
it, or an extension of evil.  Packages using child frames (like
corfu), blocks in org-mode, and other things seem to trigger unwanted
color changes, so that I can never really trust to "see" the correct
evil mode.  Plus, there is work that needs to be done with respect to
themes and changes thereof, where it's often significant whether emacs
was started with "emacsclient -a ''" or not, as the daemon mode
creates a frame, but visual settings don't have an effect until the
actually-used frame is created and visible.