atuan: add and use stumpwm
Due to an open issue in nixpkgs, pkgs.stumpwm cannot be used at the
moment, so I clone it and build it manually.

Another open problem is that the stumpwm-contrib modules are in a
separate repository, and I'm not yet sure how to make those part of my
configuration.  Maybe making them a flake input and symlinking after
the actual config would be a possibility, but it would be quite
opaque.  A git submodule might be another option here, as much as I
dislike them.