wezterm: add option to only install the config file
I might do something similar for all the things that are able to use a
GPU but have problems finding the OpenGL libraries on non-NixOS, like
kitty.

That way I can work without having to use nixGL, but have to install
these tools with the system package manager (which shouldn't be a
problem at all).