symlink-config: add module for "dynamic" configuration files
I often find myself in a position where I want to symlink a
configuration file from my dotfiles (a local checkout of *this*
repository) to the destination in my home directory. This contradicts
the idea of home-manager, where we from an immutable location within the
nix store. But since I use version control for everything, I don't
really need hm's rollback functionality for this case (or ever, really).
So I use it as a glorified GNU Stow, but with packages (and some notable
exceptions).